Feature/pet

+++ src/main/java/com/example/withdogandcat/domain/pet/PetService.java
@@ -86,9 +86,16 @@ public class PetService {
     }
 
     @Transactional
-    public BaseResponse<Void> deletePet(Long petId) {
-        Pet pet = petRepository.findById(petId).orElseThrow();
+    public BaseResponse<Void> deletePet(Long petId, User currentUser) {
+        Pet pet = petRepository.findById(petId)
+                .orElseThrow(() -> new BaseException(BaseResponseStatus.PET_NOT_FOUND));
 
+        // 현재 로그인한 사용자가 애완동물을 삭제할 권한이 있는지 확인
+        if (!pet.getUser().getUserId().equals(currentUser.getUserId())) {
+            throw new BaseException(BaseResponseStatus.ACCESS_DENIED);
+        }
+
+        // 권한이 확인되면 이미지 삭제 및 애완동물 삭제 진행
         imageS3Service.deleteImages(pet.getImages());
         petRepository.delete(pet);
         return new BaseResponse<>(BaseResponseStatus.SUCCESS, "성공", null);
